Update file structure of the project
