+++ src/components/BookStoreHeader.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useDispatch } from 'react-redux';
-import { changeFilter } from '../actions/index';
+import { changeFilter } from '../redux/index';
 import CategoryFilter from './CategoryFilter';
 
 function BookStoreHeader() {
