+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import App from './components/App';
-import store from './reducers/index';
+import store from './redux/store';
 import './stylesheet/index.scss';
 
 ReactDOM.render(
