+++ src/redux/book/bookActions.js
@@ -1,4 +1,4 @@
-import { CREATE_BOOK, REMOVE_BOOK, CHANGE_FILTER } from './actionTypes';
+import { CREATE_BOOK, REMOVE_BOOK } from './bookTypes';
 
 export const addBook = (title, category) => ({
   type: CREATE_BOOK,
@@ -12,12 +12,5 @@ export const removeBook = id => ({
   type: REMOVE_BOOK,
   payload: {
     id,
-  },
-});
-
-export const changeFilter = category => ({
-  type: CHANGE_FILTER,
-  payload: {
-    category,
   },
 });
