+++ src/redux/book/bookTypes.js
@@ -1,3 +1,2 @@
 export const CREATE_BOOK = 'CREATE_BOOK';
 export const REMOVE_BOOK = 'REMOVE_BOOK';
-export const CHANGE_FILTER = 'CHANGE_FILTER';
