+++ src/redux/filter/filterActions.js
@@ -1,0 +1,8 @@
+import { CHANGE_FILTER } from './filterTypes';
+
+export const changeFilter = category => ({
+  type: CHANGE_FILTER,
+  payload: {
+    category,
+  },
+});
