+++ src/redux/filter/filterTypes.js
@@ -1,0 +1,1 @@
+export const CHANGE_FILTER = 'CHANGE_FILTER';
